requirements.txt
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.0
pillow==12.1.0
psutil==7.2.2
reportlab==4.4.9
